fix(lens): move padiSurface output-settled watcher into the padi client kit
Relocate awaitOutputSettled (and its watcher body) from the kolu-mcp app face into packages/padi/src/watch.ts beside awaitAgentState, re-export via @kolu/padi/dial, and leave only the MCP-face glue (schema, tool, waitJson) in kolu-mcp.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-1, raised by lowy). Not pushed or merged.
